# src/main/resources/openapi/rest-article-api.yaml
openapi: 3.0.0

info:
  version: 1.0.0
  title: Simple Rest API

tags:
  - name: ArticleRest
    description: Article rest CRUD API

paths:
  /get/{id}:
    get:
      tags:
        - ArticleRest
      summary: Get one record by id
      description: Get one record by id
      operationId: get
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: ArticleVO
        '404':
          description: Request failed - No items

  /getLast:
    get:
      tags:
        - ArticleRest
      summary: Get last record
      description: Get last record
      operationId: getLastArticle
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: ArticleVO
        '404':
          description: Request failed - No items

  /getAll:
    get:
      tags:
        - ArticleRest
      summary: Get all records
      description: Get all records
      operationId: getAll
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: ArticleVO
        '404':
          description: Request failed - No items

  /getAllByParent/{id}:
    get:
      tags:
        - ArticleRest
      summary: Get all records by parent id
      description: Get all records by parent id
      operationId: getByParent
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: ArticleVO
        '404':
          description: Request failed - No items

  /add:
    post:
      tags:
        - ArticleRest
      summary: Add one message
      description: Add one message
      operationId: add
      requestBody:
        description: Data for message
        required: true
        content:
          application/json:
            schema:
              type: ArticleVO
      responses:
        '201':
          description: Successful operation'
          content:
            application/json:
              schema:
                type: ArticleVO
        '404':
          description: Request failed - No items

  /update/{id}:
    put:
      tags:
        - ArticleRest
      summary: Update one record
      description: Update one record
      operationId: update
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        description: Data to update
        required: true
        content:
          'application/json':
            schema:
              type: ArticleVO
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: ArticleVO
        '404':
          description: Request failed - No items

  /delete/{id}:
    delete:
      tags:
        - ArticleRest
      summary: Delete one record by id
      description: Delete one record by id
      operationId: delete
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Successful deleted
        '404':
          description: Request failed - No items

components:
  parameters:
    id:
      in: path
      name: id
      required: true
      description: record id
      schema:
        type: integer